+++ src/modules/bin/fs_ls.rs
@@ -8,6 +8,7 @@ use std::time::{SystemTime, UNIX_EPOCH};
 use chrono::{DateTime, NaiveDateTime};
 
 use crate::expression::FileSize;
+use crate::expression::cmd_excutor::expand_home;
 use crate::{Environment, Expression, LmError};
 
 #[derive(Default)]
@@ -39,18 +40,7 @@ pub fn parse_ls_args(args: &[Expression]) -> Result<(PathBuf, LsOptions), LmErro
                 "-u" => options.show_user = true,
                 "-m" => options.show_mode = true,
                 "-p" => options.show_path = true,
-                arg if !arg.starts_with('-') => {
-                    // dbg!("not stw -", arg);
-                    if arg.starts_with("~") {
-                        if let Some(home_dir) = dirs::home_dir() {
-                            let p = arg.replace("~", home_dir.to_string_lossy().as_ref());
-                            // dbg!(&p);
-                            path = PathBuf::from(p)
-                        }
-                    } else {
-                        path = PathBuf::from(arg)
-                    }
-                }
+                arg if !arg.starts_with('-') => path = PathBuf::from(expand_home(arg).as_ref()),
                 _ => continue,
             }
         }